added utility functions we might need
